# src/main/resources/api.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:8080
info:
  title: Restaurant Review REST API
  version: 1.0.0
  description: Spring Boot Restaurant Review REST API Documentation
  contact:
    name: Elizabeth Drew
paths:
...
  /api/v1/cuisines:
    get:
      summary: Get all cuisines
      operationId: getCuisines
      description: Lists all the available cuisines
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new cuisine
      operationId: addNewCuisine
      description: Add a new cuisine if it doesn't already exist
      security:
        - basicAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuisine'
        '400':
          description: Invalid Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Already Exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/cuisines/{name}:
    delete:
      summary: Delete a cuisine by name
      operationId: deleteCuisine
      description: Delete an unused cuisine
      security:
        - basicAuth: [ ]
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the cuisine to delete
          schema:
            type: string
      responses:
        '204':
          description: Successfully Deleted
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Referenced By Restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/claims/pending:
    get:
      summary: Get all pending claims for restaurant ownership
      operationId: getPendingClaims
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: List of all pending claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClaimStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/claims/{claimId}/accept:
    post:
      summary: Admin accepts a claim
      description: Admin accepts a claim for restaurant ownership
      operationId: acceptClaim
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      parameters:
        - name: claimId
          in: path
          description: ID of the claim to be accepted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Claim accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimStatus'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/claims/{claimId}/reject:
    post:
      summary: Admin rejects a claim
      description: Admin rejects a claim for restaurant ownership
      operationId: rejectClaim
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      parameters:
        - name: claimId
          in: path
          description: ID of the claim to be rejected
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Claim rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimStatus'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/users/pending:
    get:
      summary: Get all pending user requests for admin account
      operationId: getPendingAdminAccounts
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: List of all pending admin requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/users/{requestId}/accept:
    post:
      description: Admin accepts a user request for admin account
      operationId: acceptAdminAccount
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      parameters:
        - name: requestId
          in: path
          description: ID of the request to be accepted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Admin account accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatus'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/users/{requestId}/reject:
    post:
      description: Admin rejects an admin account request
      operationId: rejectAdminAccount
      tags:
        - Admin
      security:
        - basicAuth: [ ]
      parameters:
        - name: requestId
          in: path
          description: ID of the request to be rejected
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Admin account rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatus'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Cuisine:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message
    ClaimStatus:
      type: object
      properties:
        status:
          type: string
          description: The claim status (e.g., pending, accepted, rejected)
        claimant:
          type: integer
          format: int64
          description: ID of the user who has made the claim
        restaurant:
          type: integer
          format: int64
          description: ID of the restaurant the claim relates to
        claim_id:
          type: integer
          format: int64
          description: ID of the claim
        reason:
          type: string
          description: Reason or explanation provided by the user for the claim
...
    AdminStatus:
      type: object
      properties:
        status:
          type: string
          description: The claim status (e.g., pending, accepted, rejected)
        user_id:
          type: integer
          format: int64
          description: ID of the user who has made the claim
        request_id:
          type: integer
          format: int64
          description: ID of the request